don't catch timeout errors
We want to make sure (using whichever strategy) that
the application we wait for is running. If we catch the error we hide the
fact that the app has not started properly. If a developer don't want to make
sure the app is running, then they should not use `join` method :)